// src/components/dashboard/support-desk/types/top-performer-help-agent.types.ts
/**
 * TopPerformerHelpAgentTypes
 * -----------------------------------------------------------------------------
 * This file defines the TypeScript interface for the Top Performer Help Agent
 * data structure used in the `TopPerformerHelpAgent` component.
 */

export interface TopPerformerHelpAgentItem {
  id: number
  image: string
  name: string
  AgentId: string
  Average: string
}